Fantom 1.0.71 files

+++ com.xored.f4.fantom/fantom/src/sys/java/fan/sys/LocalFile.java
@@ -69,17 +69,28 @@ public class LocalFile
   {
     String path = uri.pathStr();
     int len = path.length();
-    StringBuilder s = null;
+
+    // check for escapes
+    boolean hasEsc = false;
+    for (int i=0; i<len; ++i)
+      if (path.charAt(i) == '\\') { hasEsc = true; break; }
+    if (!hasEsc) return path;
+
+    // normalize
+    StringBuilder s = new StringBuilder(len);
     for (int i=0; i<len; ++i)
     {
       int c = path.charAt(i);
       if (c == '\\')
       {
-        if (s == null) { s = new StringBuilder(); s.append(path, 0, i); }
+        i++;
+        if (i>=len) throw ArgErr.make("Invalid Uri esc: " + path);
+        c = path.charAt(i);
+        if (c == '.') throw ArgErr.make("Invalid Uri esc: " + path);
       }
-      else if (s != null) s.append((char)c);
+      s.append((char)c);
     }
-    return s == null ? path : s.toString();
+    return s.toString();
   }
 
   public static String fileNameToUriName(String name)
@@ -211,6 +222,8 @@ public class LocalFile
     file.setLastModified(time.toJava());
   }
 
+  public boolean isHidden() { return file.isHidden(); }
+
   public String osPath()
   {
     return file.getPath();
@@ -279,12 +292,44 @@ public class LocalFile
   {
     try
     {
-      return new LocalFileStore(this.file);
+      return new LocalFileStore(java.nio.file.Files.getFileStore(this.file.toPath()));
     }
     catch (java.io.IOException e)
     {
       throw IOErr.make(e);
     }
+  }
+
+//////////////////////////////////////////////////////////////////////////
+// Trap
+//////////////////////////////////////////////////////////////////////////
+
+  public Object trap(String name, List args)
+  {
+    // undocumented local file access
+    Object arg = args == null ? null : args.first();
+    if (name.equals("readable"))   return trapReadable(arg);
+    if (name.equals("writable"))   return trapWritable(arg);
+    if (name.equals("executable")) return trapExecutable(arg);
+    return super.trap(name, args);
+  }
+
+  private Object trapReadable(Object arg)
+  {
+    if (arg == null) return this.file.canRead();
+    else return this.file.setReadable((Boolean)arg);
+  }
+
+  private Object trapWritable(Object arg)
+  {
+    if (arg == null) return this.file.canWrite();
+    else return this.file.setWritable((Boolean)arg);
+  }
+
+  private Object trapExecutable(Object arg)
+  {
+    if (arg == null) return this.file.canExecute();
+    else return this.file.setExecutable((Boolean)arg);
   }
 
 //////////////////////////////////////////////////////////////////////////
@@ -372,17 +417,24 @@ public class LocalFile
 
   public void delete()
   {
-    if (!exists()) return;
-
-    if (file.isDirectory())
+    if (exists() && file.isDirectory())
     {
       List kids = list();
       for (int i=0; i<kids.sz(); ++i)
         ((File)kids.get(i)).delete();
     }
 
-    if (!file.delete())
-      throw IOErr.make("Cannot delete: " + file);
+    try
+    {
+      // java.io.File has some issues on macOS (and Linux?) with
+      // broken symlinks; and will report they do not exist; use
+      // Files.deleteIfExists to cleanup properly
+      java.nio.file.Files.deleteIfExists(file.toPath());
+    }
+    catch (java.io.IOException err)
+    {
+      throw IOErr.make("Cannot delete: " + file, err);
+    }
   }
 
   public File deleteOnExit()
